requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.22.0
requests>=2.28.0
openai-agents==0.0.4